TestConnections/requirements.txt
dronekit==2.9.2
future==1.0.0
lxml==5.3.1
MAVProxy==1.8.71
monotonic==1.6
numpy==2.2.4
pymavlink==2.4.43
pynmeagps==1.0.46
pyserial==3.5
